+++ CHANGELOG.md
@@ -42,8 +42,12 @@ NOTE: Add new changes BELOW THIS COMMENT.
 
 ### Fixed
 
+- Updater installations when the working directory differs from the executable
+  directory ([#3286]).
+
 - Blocked requests without an EDNS(0) OPT record ([#8183]).
 
+[#3286]: https://github.com/AdguardTeam/AdGuardHome/issues/3286
 [#8183]: https://github.com/AdguardTeam/AdGuardHome/issues/8183
 
 <!--